pipeline/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
